F_T30603 move DocumentBundle Tools to core
- adjut use statements
