+++ demosplan/DemosPlanCoreBundle/Tools/ServiceImporter.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanDocumentBundle\Tools;
+namespace demosplan\DemosPlanCoreBundle\Tools;
 
 use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
 use DemosEurope\DemosplanAddon\Contracts\MessageBagInterface;
@@ -33,6 +33,7 @@ use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Routing\RouterInterface;
+use function demosplan\DemosPlanDocumentBundle\Tools\count;
 
 /**
  * Import von Planunterlagen-Absaetzen.
